requirements.txt
python==3.7.13
mysql-connector-python==8.0.18
xlrd==2.0.1
openpyxl==3.0.9
flask==2.1.2